Cleanup specs
