+++ PistachioTests/DictionaryAdapterSpec.swift
@@ -1,0 +1,26 @@
+//  Copyright (c) 2015 Felix Jendrusch. All rights reserved.
+
+import Quick
+import Nimble
+
+import Pistachio
+
+class DictionaryAdapterSpec: QuickSpec {
+    override func spec() {
+        describe("A DictionaryAdapter") {
+            let adapter = NodeAdapters.anyObject
+
+            it("should transform a value") {
+                let result = adapter.transform(Node(children: [ Node(children: []) ]))
+
+                expect(((result.value as? [String: AnyObject])?["children"] as? [AnyObject])?.count).to(equal(1))
+            }
+
+            it("should reverse transform a value") {
+                let result = adapter.reverseTransform([ "children": [ [ "children": [] ] ] ])
+
+                expect(result.value?.children.count).to(equal(1))
+            }
+        }
+    }
+}
